add Interpreter::reset clearing the value stack and rewinding ip, call reset at start of run so a compiled interpreter can be re-executed repeatedly (failed runs no longer poison the next one), add tests for compile-once-run-many-times, explicit reset, and clean restart after a failed run, drop redundant f64 casts in lib and cli Math fixtures, use is_none in reader test

+++ examples/cli.rs
@@ -27,7 +27,7 @@ impl Callable<MethodId> for Math {
         match (op, args) {
             (MethodId::Sum, &[l, r]) => l + r,
             (MethodId::Sub, &[l, r]) => l - r,
-            (MethodId::Foo, &[l]) => l as f64,
+            (MethodId::Foo, &[l]) => l,
             _ => todo!(),
         }
     }
